+++ src/lib/mixer_module/mixer_module_tests.cpp
@@ -87,7 +87,7 @@ public:
 		was_scheduled = true;
 	}
 
-	bool updateOutputs(bool stop_motors, int16_t outputs_[MAX_ACTUATORS],
+	bool updateOutputs(bool stop_motors, float outputs_[MAX_ACTUATORS],
 			   unsigned num_outputs_, unsigned num_control_groups_updated) override
 	{
 		memcpy(outputs, outputs_, sizeof(outputs));
